reorganize browser package folders
