+++ packages/browser/src/index.ts
@@ -1,11 +1,11 @@
 // public API
-export { BrowserContext, register } from "./managers/ContextManager";
-export { Page } from "./managers/PageManager";
+export { BrowserContext, register } from "./ContextManager";
+export { Page } from "./page/PageManager";
 // TODO move to jest-playwright
 export { launch, LaunchResult } from "./launch";
 
 // internal API
 export * from "./actions";
 export * from "./keyboard";
-export { ContextManager } from "./managers/ContextManager";
-export { PageManager } from "./managers/PageManager";
+export { ContextManager } from "./ContextManager";
+export { PageManager } from "./page/PageManager";
