+++ packages/browser/src/page/buildScript.ts
