+++ packages/browser/tests/page/captureLogs.test.ts
